AImodel/requirements.txt
numpy==1.26.4
werkzeug==2.0.3
flask==2.0.1
tensorflow==2.17.1
Pillow==10.0.0
gunicorn==20.1.0